Unit 1: Assignment 2
Introduction
This assignment lets you practice with writing a program that accepts user input and performs some simple calculations. You will be using some of the arithmetic operators discussed in the text.

Directions
Write a Python program that prompts the user to enter a number of quarts, and then calculates the number of barrels and remaining quarts, as well as the number of ounces, and then prints the results.

Notes
There are 32 ounces in a quart and 168 quarts in a barrel. Your program must declare and use several variables, like for the number of quarts and number of ounces. You should reread the text and look at the examples to figure out the arithmetic by hand before you start to code.

You should choose appropriate names based on your reading.

Your output should look like the examples. Be sure to print text and numbers without extra spaces.